Prediksi-Harga-Rumah/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
pyyaml>=6.0
Pillow>=10.0.0

